# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart Irrigation System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pumps/{id}/control:
    post:
      operationId: controlPump
      tags: [pumps]
      summary: Control pump state (on/off/override)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PumpControlBody"
      responses:
        "200":
          description: Updated pump state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pump"
...
  /sensors/history:
    get:
      operationId: getSensorHistory
      tags: [sensors]
      summary: Get historical sensor data
      parameters:
        - name: hours
          in: query
          schema:
            type: integer
            default: 24
        - name: zoneId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Sensor history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SensorHistoryRecord"
  /esp32/ingest:
    post:
      operationId: ingestEsp32Data
      tags: [esp32]
      summary: Ingest sensor data from ESP32
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Esp32IngestBody"
      responses:
        "200":
          description: Data accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestResponse"
...
  /irrigation/logs:
    get:
      operationId: getIrrigationLogs
      tags: [irrigation]
      summary: Get irrigation action logs
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: zoneId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Irrigation logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IrrigationLog"
    post:
      operationId: createIrrigationLog
      tags: [irrigation]
      summary: Create manual irrigation log entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIrrigationLogBody"
      responses:
        "201":
          description: Log created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IrrigationLog"
  /alerts:
    get:
      operationId: getAlerts
      tags: [alerts]
      summary: Get active alerts
      parameters:
        - name: acknowledged
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Alerts list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
  /alerts/{id}/acknowledge:
    patch:
      operationId: acknowledgeAlert
      tags: [alerts]
      summary: Acknowledge or dismiss an alert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
...
  /crops/{id}:
    get:
      operationId: getCrop
      tags: [crops]
      summary: Get crop details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Crop details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Crop"
...
components:
  schemas:
...
    Pump:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        zoneId:
          type: integer
        zoneName:
          type: string
        status:
          type: string
          enum: [on, off, override]
        isManualOverride:
          type: boolean
        driverChannel:
          type: string
        runtimeToday:
          type: number
        lastToggled:
          type: string
          nullable: true
      required:
        - id
        - name
        - zoneId
        - zoneName
        - status
        - isManualOverride
        - driverChannel
        - runtimeToday
    PumpControlBody:
      type: object
      properties:
        action:
          type: string
          enum: [on, off, auto]
        isOverride:
          type: boolean
      required: [action]
...
    SensorHistoryRecord:
      type: object
      properties:
        id:
          type: integer
        zoneId:
          type: integer
          nullable: true
        moisture:
          type: number
          nullable: true
        temperature:
          type: number
          nullable: true
        humidity:
          type: number
          nullable: true
        tankLevelPercent:
          type: number
          nullable: true
        waterLoggingDetected:
          type: boolean
          nullable: true
        recordedAt:
          type: string
      required: [id, recordedAt]
    Esp32IngestBody:
      type: object
      properties:
        moisture1:
          type: number
        moisture2:
          type: number
        moisture3:
          type: number
        moisture4:
          type: number
        temperature:
          type: number
        humidity:
          type: number
        tankLevel:
          type: number
        waterLogging:
          type: boolean
        deviceId:
          type: string
      required:
        - moisture1
        - moisture2
        - moisture3
        - moisture4
        - temperature
        - humidity
        - tankLevel
        - waterLogging
    IngestResponse:
      type: object
      properties:
        success:
          type: boolean
        pumpCommands:
          type: array
          items:
            type: object
            properties:
              pumpId:
                type: integer
              action:
                type: string
            required: [pumpId, action]
        message:
          type: string
      required: [success, pumpCommands, message]
...
    IrrigationLog:
      type: object
      properties:
        id:
          type: integer
        zoneId:
          type: integer
        zoneName:
          type: string
        pumpId:
          type: integer
        action:
          type: string
          enum: [started, stopped]
        trigger:
          type: string
          enum: [auto, manual, override, ai]
        durationSeconds:
          type: integer
          nullable: true
        soilMoistureBefore:
          type: number
          nullable: true
        soilMoistureAfter:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - zoneId
        - zoneName
        - pumpId
        - action
        - trigger
        - createdAt
    CreateIrrigationLogBody:
      type: object
      properties:
        zoneId:
          type: integer
        pumpId:
          type: integer
        action:
          type: string
          enum: [started, stopped]
        trigger:
          type: string
          enum: [auto, manual, override, ai]
        durationSeconds:
          type: integer
        soilMoistureBefore:
          type: number
        soilMoistureAfter:
          type: number
        notes:
          type: string
      required: [zoneId, pumpId, action, trigger]
    Alert:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [tank_empty, water_logging, sensor_anomaly, pump_failure, low_moisture, high_moisture]
        severity:
          type: string
          enum: [info, warning, critical]
        message:
          type: string
        zoneId:
          type: integer
          nullable: true
        acknowledged:
          type: boolean
        createdAt:
          type: string
        acknowledgedAt:
          type: string
          nullable: true
      required:
        - id
        - type
        - severity
        - message
        - acknowledged
        - createdAt
    Crop:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        idealMoistureMin:
          type: number
        idealMoistureMax:
          type: number
        waterRequirementMm:
          type: number
        growthStages:
          type: array
          items:
            type: string
        currentStage:
          type: string
          nullable: true
        description:
          type: string
      required:
        - id
        - name
        - idealMoistureMin
        - idealMoistureMax
        - waterRequirementMm
        - growthStages
        - description
...